+++ src/index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" />
     <title>Blase TV</title>
 </head>
-<body onload="renderListOfItems()">
+<body id="body">
 
 <!-- Header Section -->
 <header id="nav-bar">
