+++ src/index.js
@@ -1,15 +1,5 @@
 import './index.css';
+import listOfMovies from './modules/homepage.js';
 
-const hearts = document.querySelectorAll('.heart');
-console.dir(hearts);
-hearts.forEach((heart) => {
-  heart.addEventListener('click', () => {
-    if (heart.classList.contains('heart-uncheck')) {
-      heart.classList.remove('heart-uncheck');
-      heart.classList.add('heart-check');
-    } else {
-      heart.classList.remove('heart-check');
-      heart.classList.add('heart-uncheck');
-    }
-  });
-});
+const body = document.getElementById('body');
+body.onload = () => { window.listOfMovies = listOfMovies(); };
