+++ src/modules/homepage.js
@@ -1,11 +1,24 @@
-import {getMovieCard} from './movieCard.js';
+import getMovieCard from './movieCard.js';
 
-window.renderListOfItems = async () => {
+const listOfMovies = async () => {
   const data = await fetch('https://api.tvmaze.com/shows/1/episodes?api_key=0hDJ3ZtNQRDdBTeYbSAz7xickoKLyd6A');
   const shows = await data.json();
-  let cards = ``;
+  let cards = '';
   shows.forEach((show) => {
     cards += getMovieCard(show.image.original, show.name);
-  })
-  document.querySelector('.card__section').setHTML(cards);
+  });
+  document.querySelector('.card__section').innerHTML = cards;
+  const hearts = document.querySelectorAll('.heart');
+  hearts.forEach((heart) => {
+    heart.addEventListener('click', () => {
+      if (heart.classList.contains('heart-uncheck')) {
+        heart.classList.remove('heart-uncheck');
+        heart.classList.add('heart-check');
+      } else {
+        heart.classList.remove('heart-check');
+        heart.classList.add('heart-uncheck');
+      }
+    });
+  });
 };
+export default listOfMovies;
