+++ src/modules/movieCard.js
@@ -1,9 +1,8 @@
-export const getMovieCard = (poster, title) => {
-  return `<div class="card">
+const getMovieCard = (poster, title) => `<div class="card">
                 <img class="poster" src="${poster}" alt="Poster">
                 <div class="card__header">
                     <p class="card__heading">${title}</p>
-                    <i class="fa-solid fa-heart heart heart-uncheck"></i>
+                    <i class="fa-solid fa-heart heart heart-uncheck" ></i>
                     <p class="card__total-likes">2 Likes</p>
                 </div>
                 <div class="card__footer">
@@ -11,4 +10,5 @@ export const getMovieCard = (poster, title) => {
                     <div class="card__reservation card-btn">Reservations</div>
                 </div>
             </div>`;
-}
+
+export default getMovieCard;
